+++ serving/vllm/shim/__init__.py
@@ -1,0 +1,1 @@
+"""Vulcan vLLM contract shim package."""